requirements.txt
streamlit==1.44.1
pydantic==2.11.1
docxtpl==0.20.0
python-docx==1.1.2
pywin32==308
